refactor(team-builder): split IdentityLaneReal into three layout components
Replace the {isCompact ? compactBranch : heroBranch} JSX duality in
IdentityLaneReal with three distinct layout components, each rendering
one static shape composed of the atomic cells extracted in steps 2–5:

- IdentitySingleRow — slot >=1240, renders the existing compact branch
- IdentityMidStack — slot <1240, renders the existing hero branch
- IdentityVertical — skeleton for the new 2x3-vertical / 3x2-vertical
  modes coming in step 8 (not yet routed by the dispatcher)

The new IdentityLane dispatcher reads useContainerCompact + the
TeamLayoutContext mode and picks one layout to mount. Each layout
component owns its own useIdentityState() call and its own
SpeciesPickerDialog mount.

No visual change. Each new layout still renders the existing CSS
classes (.identCompact / .identHero), so @container queries and
[data-layout] overrides continue to produce identical output. The CSS
cutover to per-component scoped classes happens in step 7.

Layout files contain zero useState calls beyond the speciesOpen dialog
flag, zero picker imports, zero direct onUpdate handler definitions --
pure markup composition.

A shared IdentityLayoutProps interface lives in identity-layout-props.ts
so all three layouts reference the same type without redeclaration.

TeamLayoutMode is widened to string in the dispatcher before the
"-vertical" suffix check, since those mode values don't exist in the
type union yet (step 8 adds them) -- avoids narrowing-to-never TS error.

apps/web/src/components/team-builder/v2/lanes/identity-lane.tsx is now
a thin re-export from ./identity/. The old IdentityLaneReal body is
gone. Step 9 deletes this stub once active-row.tsx updates its import.
