+++ apps/web/src/components/team-builder/v2/lanes/identity/index.ts
@@ -1,0 +1,1 @@
+export { IdentityLane } from "./identity-lane";
